:recycle:(chore) separate infra files from domain
