+++ server/src/modules/appointments/infra/typeorm/entities/Appointment.ts
@@ -8,8 +8,8 @@ import {
   JoinColumn,
 } from 'typeorm';
 
-import { APPOINTMENTS_TABLE_NAME } from '../../../shared/database/migrations/1600388380111-CreateAppointments';
-import User from '../../users/entities/User';
+import { APPOINTMENTS_TABLE_NAME } from '../../../../../shared/infra/typeorm/migrations/1600388380111-CreateAppointments';
+import User from '../../../../users/infra/typeorm/entities/User';
 
 @Entity(APPOINTMENTS_TABLE_NAME)
 class Appointment {
