+++ server/src/modules/users/repositories/UsersRepository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import User from '../entities/User';
+import User from '../infra/typeorm/entities/User';
 
 @EntityRepository(User)
 class UsersRepository extends Repository<User> {
