+++ server/src/shared/infra/http/middlewares/ensureAuthentication.ts
@@ -1,8 +1,8 @@
 import { StatusCodes } from 'http-status-codes';
 import { verify } from 'jsonwebtoken';
-import { ExpressMiddleware } from '../../@types/middleware';
-import authConfig from '../../config/auth-config';
-import AppException from '../exceptions/AppException';
+import { ExpressMiddleware } from '../../../../@types/middleware';
+import authConfig from '../../../../config/auth-config';
+import AppException from '../../../exceptions/AppException';
 
 interface TokenPayload {
   iat: number;
