+++ server/src/shared/infra/http/middlewares/generalException.ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
-import { ExpressErrorMiddleware } from '../../@types/middleware';
-import AppException from '../exceptions/AppException';
+import { ExpressErrorMiddleware } from '../../../../@types/middleware';
+import AppException from '../../../exceptions/AppException';
 
 const generalException: ExpressErrorMiddleware = (
   err,
