+++ server/src/shared/infra/http/routes/v1/appointments.routes.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 import { parseISO } from 'date-fns';
 import { getCustomRepository } from 'typeorm';
-import AppointmentsRepository from '../../../modules/appointments/repositories/AppointmentsRepository';
-import CreateAppointmentService from '../../../modules/users/services/CreateAppointmentService';
+import AppointmentsRepository from '../../../../../modules/appointments/repositories/AppointmentsRepository';
+import CreateAppointmentService from '../../../../../modules/users/services/CreateAppointmentService';
 import ensureAuthentication from '../../middlewares/ensureAuthentication';
 
 const routes = Router();
