+++ server/src/shared/infra/http/routes/v1/sessions.routes.ts
@@ -1,6 +1,6 @@
 import { classToPlain } from 'class-transformer';
 import { Router } from 'express';
-import AuthenticateUserService from '../../../modules/appointments/services/AuthenticateUserService';
+import AuthenticateUserService from '../../../../../modules/appointments/services/AuthenticateUserService';
 
 const routes = Router();
