+++ server/src/shared/infra/http/routes/v1/users.routes.ts
@@ -1,10 +1,10 @@
 import { classToPlain } from 'class-transformer';
 import { Router } from 'express';
 import multer from 'multer';
-import uploadConfig from '../../../config/upload-config';
+import uploadConfig from '../../../../../config/upload-config';
 import ensureAuthentication from '../../middlewares/ensureAuthentication';
-import CreateUserService from '../../../modules/users/services/CreateUserService';
-import UpdateUserAvatarService from '../../../modules/users/services/UpdateUserAvatarService';
+import CreateUserService from '../../../../../modules/users/services/CreateUserService';
+import UpdateUserAvatarService from '../../../../../modules/users/services/UpdateUserAvatarService';
 
 const routes = Router();
 const upload = multer(uploadConfig);
